+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 * Add config option `indent`
 * Show metadata with hledger output
+* Support setting auxiliary dates and posting dates from metadata ([issue #14](https://github.com/beancount/beancount2ledger/issues/14))
 
 ## 1.2 (2020-07-25)